Correcting invoice post validation and modularizing validation
