+++ api/src/services/invoices/index.js
@@ -1,4 +1,3 @@
-import { runInNewContext } from 'vm';
 import ResponseList from '../../utils/response';
 import { limitSettings } from '../../config/config';
 import models from '../../models';
